No disk

+++ ga2/gaUtils/gaDeamon.py
@@ -8,31 +8,36 @@ import time
 import random
 import pickle
 
+import ga2
 import ga2.gaUtils.AgentClass as AgentClass
 GA_UTIL_DIR = os.path.dirname(os.path.realpath(__file__))
 
 
 # current functions included are as follows:
 #
-# ** uses lock
-# *
-# 
-# 1 ) ** getNewSessionID()
-# 2 ) isLocked()
-# 3 ) lock()
-# 4 ) unlock()
-# 5 ) getCurrGen()
-# 6 ) setCuttGen()
-# 7 ) ** storeAgent()
-# 8 ) generateAgentID()
-# 9 ) createAgent()
-# 10) ** deleteAgent()
-# 11) getAgent()
-# 12) ** setSession()
-# 13) getSession()
-# 14) asexualRep()
-# 15) mutateAgentObj()
-# 16) sexualRep()
+# **   uses lock
+# %%   uses Disk
+# $$   legacy
+# @@   changable
+# --   discontinued
+# ^^   finished
+#
+# 1 ) ** %%       ^^  getNewSessionID()
+# 2 )    %%    @@  ^  isLocked()
+# 3 )    %%    @@  ^  lock()
+# 4 )    %%    @@  ^  unlock()
+# 5 )    %%    @@ ^^  getCurrGen()
+# 6 )    %%    @@ ^^  setCurrGen()
+# 7 ) ** %%    @@ ^^  storeAgent()
+# 8 )    %%    @@ ^^  generateAgentID()
+# 9 )             ^^  createAgent()
+# 10) ** %%    @@ ^^  deleteAgent()
+# 11)    %%    @@ ^^  getAgent()
+# 12) -- -- -- -- --  setSession()
+# 13) -- -- -- -- --  getSession()
+# 14)             ^^  asexualRep()
+# 15)             ^^  mutateAgentObj()
+# 16)             ^^  sexualRep()
 
 
 def mutateAgentObj(agentObj, sess):
@@ -117,52 +122,52 @@ def sexualRep(sess, parent1Obj, parent2Obj):
     return(childObj)
 
 
-def getSession(sessID):
-    """
-    INPUT       : sessID
-    OUTPUT      : Agent object
+# def getSession(sessID):
+#     """
+#     INPUT       : sessID
+#     OUTPUT      : Agent object
 
-    DESCRIPTION : Creates and stores the newly created AgentDna
-                  object in that particular session's directory
-    """
-    try:
+#     DESCRIPTION : Creates and stores the newly created AgentDna
+#                   object in that particular session's directory
+#     """
+#     try:
     
-        cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)
-                    +"/smtf/SESS.smtf", "rb")
-        currImg = pickle.load(cgfp)
-        cgfp.close()
+#         cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)
+#                     +"/smtf/SESS.smtf", "rb")
+#         currImg = pickle.load(cgfp)
+#         cgfp.close()
     
-    except Exception:
-        print("error in getting the session {}".format(sessID))
-        currImg = None
+#     except Exception:
+#         print("error in getting the session {}".format(sessID))
+#         currImg = None
         
-    print(currImg.agentCount)
-    return(currImg)
+#     print(currImg.agentCount)
+#     return(currImg)
 
 
-def setSession(sessionObj):
-    """
-    INPUT       : sessionObj
-    OUTPUT      : 1 on success 0 on failure
+# def setSession(sessionObj):
+#     """
+#     INPUT       : sessionObj
+#     OUTPUT      : 1 on success 0 on failure
 
-    DESCRIPTION : Creates and stores the newly created AgentDna
-                  object in that particular session's directory
-    """
+#     DESCRIPTION : Creates and stores the newly created AgentDna
+#                   object in that particular session's directory
+#     """
     
-    lock(sessionObj.sessID)
-    #try:
-    tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessionObj.sessID)
-                +"/smtf/SESS.smtf", "wb")
-    pickle.dump(sessionObj, tpfp)
-    tpfp.close()
-    '''
-    except Exception:
-        print("error in store of session, couldnt wb")
-        return(0)
-    '''
-    unlock(sessionObj.sessID)
+#     lock(sessionObj.sessID)
+#     #try:
+#     tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessionObj.sessID)
+#                 +"/smtf/SESS.smtf", "wb")
+#     pickle.dump(sessionObj, tpfp)
+#     tpfp.close()
+#     '''
+#     except Exception:
+#         print("error in store of session, couldnt wb")
+#         return(0)
+#     '''
+#     unlock(sessionObj.sessID)
     
-    return(1)
+#     return(1)
 
 
 def getNewSessionID():
@@ -174,7 +179,6 @@ def getNewSessionID():
                   first session
     """
 
-    lock()
     try:
         
         sessFp = open(GA_UTIL_DIR+"/utilFiles/SESSIONNEXT.smtf", "r")
@@ -192,14 +196,11 @@ def getNewSessionID():
 
     except Exception:
         print("Error updating the SESSIONNEXT.smtf file")
-
-    unlock()
-    lock(sessID)
-    unlock(sessID)
+        
     return(sessID)
 
 
-def isLocked(sessID=None):
+def isLocked(sess):
     """
     INPUT       : sessID = None
     OUTPUT      : current status of the lock based on 
@@ -208,25 +209,29 @@ def isLocked(sessID=None):
     DESCRIPTION : checks the locks, based on sessID
     """
 
-    if(sessID is not None):
-        smtf = "tmp" + str(sessID) + "/LOCK.smtf"
-    else:
-        smtf = "SESSIONLOCK.smtf" 
+    if(sess.mode == 'SAFE'):
+        sessID = 10 # temposrary assignment
+        if(sessID is not None):
+            smtf = "tmp" + str(sessID) + "/LOCK.smtf"
+        else:
+            smtf = "SESSIONLOCK.smtf" 
     
-    try:
-        if(not os.path.isdir(GA_UTIL_DIR+"/utilFiles")):
-            os.mkdir(GA_UTIL_DIR+"/utilFiles")
-        lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "r")
-        lckpos = int(lckfp.readline())
-        lckfp.close()
+        try:
+            
+            if(not os.path.isdir(GA_UTIL_DIR+"/utilFiles")):
+                os.mkdir(GA_UTIL_DIR+"/utilFiles")
+            lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "r")
+            lckpos = int(lckfp.readline())
+            lckfp.close()
 
-    except Exception:
-        lckpos = 0
+        except Exception:
+            lckpos = 0
+            
+        return(lckpos)
+    
 
-    return(lckpos)
-
-
-def lock(sessID=None):
+    
+def lock(sess):
     """
     INPUT       : sessID = None
     OUTPUT      : returns 1 if locked, 0 on error
@@ -235,31 +240,34 @@ def lock(sessID=None):
                   else , does it for the entire session
     """
 
-    while(isLocked(sessID)):
-        time.sleep(0.5)
-
-    if(sessID is not None):
-        smtf = "tmp" + str(sessID) + "/smtf/LOCK.smtf"
-    else:
-        smtf = "SESSIONLOCK.smtf" 
-        
-    try:
+    if(sess.mode == 'SAFE'):
+        sessID = sess.sessID # temporary assignment
+        while(isLocked(sess)):
+            time.sleep(0.5)
+            
+        if(sessID is not None):
+            smtf = "tmp" + str(sessID) + "/smtf/LOCK.smtf"
+        else:
+            smtf = "SESSIONLOCK.smtf" 
+            
+        try:
+              
+            lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "w")
+            lckfp.close()
+            
+        except Exception:
+            os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID))
+            os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID)+"/smtf")
+            os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID)+"/dnaPool")
+            
+        lckval = 1
         lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "w")
+        lckfp.write(str(lckval))
         lckfp.close()
-    
-    except Exception:
-        os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID))
-        os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID)+"/smtf")
-        os.mkdir(GA_UTIL_DIR+"/utilFiles/"+"tmp"+str(sessID)+"/dnaPool")
-        
-    lckval = 1
-    lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "w")
-    lckfp.write(str(lckval))
-    lckfp.close()
-    return(lckval)
+        return(lckval)
 
 
-def unlock(sessID=None):
+def unlock(sess):
     """
     INPUT       : sessID = None
     OUTPUT      : 1 on successful unlocking 0 on unsuccessful
@@ -268,24 +276,26 @@ def unlock(sessID=None):
                   else , does it for the entire session
     """
 
-    if(sessID is not None):
-        smtf = "tmp" + str(sessID) + "/smtf/LOCK.smtf"
-    else:
-        smtf = "SESSIONLOCK.smtf"
-
-    try:
-        lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "w")
-        lckfp.write(str(0))
-        lckfp.close()
+    if(sess.mode == 'SAFE'):
+        sessID = sess.sessID
+        if(sessID is not None):
+            smtf = "tmp" + str(sessID) + "/smtf/LOCK.smtf"
+        else:
+            smtf = "SESSIONLOCK.smtf"
+        try:
+                
+            lckfp = open(GA_UTIL_DIR+"/utilFiles/"+smtf, "w")
+            lckfp.write(str(0))
+            lckfp.close()
+            
+        except Exception:
+            print("Couldnt unlock file !")
+            return(0)
         
-    except Exception:
-        print("Couldnt unlock file !")
-        return(0)
+        return(1)
     
-    return(1)
 
-
-def getCurrGen(sessID):
+def getCurrGen(sess):
     """
     INPUT       : sessID
     OUTPUT      : Agent object
@@ -294,17 +304,21 @@ def getCurrGen(sessID):
                   object in that particular session's directory
     """
     try:
-        cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)+"/smtf/VALID.smtf", "rb")
-        currImg = pickle.load(cgfp)
-        cgfp.close()
+        if(sess.mode=='SAFE'):
+            cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sess.sessID)+"/smtf/VALID.smtf", "rb")
+            currImg = pickle.load(cgfp)
+            cgfp.close()
+        else:
+            currImg = sess.currGen
+    
     except Exception:
         #print("error in getting the curr gen")
         currImg = set()
-
+    
     return(currImg)
 
 
-def setCurrGen(sessID, agentList):
+def setCurrGen(sess, agentList):
     """
     INPUT       : None
     OUTPUT      : Agent object
@@ -312,17 +326,22 @@ def setCurrGen(sessID, agentList):
     DESCRIPTION : Creates and stores the newly created AgentDna
                   object in that particular session's directory
     """
-    try:
-        cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)+"/smtf/VALID.smtf", "wb")
-        pickle.dump(agentList, cgfp)
-        cgfp.close()
+    if(1):
+        if(sess.mode == 'SAFE'):
+            cgfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sess.sessID)+"/smtf/VALID.smtf", "wb")
+            pickle.dump(agentList, cgfp)
+            cgfp.close()
+        else:
+            sess.currGen = agentList
+    '''
     except Exception:
         print("error in setting curr gen")
         return(0)
+    '''
     return(1)
 
 
-def storeAgent(agentObj):
+def storeAgent(sess, agentObj):
     """
     INPUT       : None
     OUTPUT      : Agent object
@@ -331,24 +350,29 @@ def storeAgent(agentObj):
                   object in that particular session's directory
     """
     
-    currAgents = getCurrGen(agentObj.sessID)
-    lock(agentObj.sessID)
-    try:
-        tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(agentObj.sessID)+"/dnaPool/dna"
-                    +str(agentObj.agentID)+".dna", "wb")
-        pickle.dump(agentObj, tpfp)
-        tpfp.close()
-        currAgents.add(agentObj.agentID)
+    currAgents = getCurrGen(sess)
+    lock(sess)
+    if(1):
+        if(sess.mode == 'SAFE'):
+            tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(agentObj.sessID)+"/dnaPool/dna"
+                        +str(agentObj.agentID)+".dna", "wb")
+            pickle.dump(agentObj, tpfp)
+            tpfp.close()
+            currAgents.add(agentObj.agentID)
+        else:
+            sess.agentBasket[agentObj.agentID] = agentObj
+            currAgents.add(agentObj.agentID)
+    '''
     except Exception:
         print("error in store agent, couldnt wb")
         return(0)
-    
-    setCurrGen(agentObj.sessID, currAgents)
-    unlock(agentObj.sessID)
+    '''
+    setCurrGen( sess, currAgents)
+    unlock(sess)
     return(agentObj.agentID)
 
 
-def getAgent(sessID, agentID):
+def getAgent(sess, agentID):
     """
     INPUT       : None
     OUTPUT      : Agent object
@@ -357,20 +381,23 @@ def getAgent(sessID, agentID):
                   object in that particular session's directory
     """
 
-    try:
-        tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)+"/dnaPool/dna"
-                    +str(agentID)+".dna", "rb")
-        agentObj = pickle.load(tpfp)
-        tpfp.close()
-    
+    if(1):
+        if(sess.mode == 'SAFE'):
+            tpfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sess.sessID)+"/dnaPool/dna"
+                        +str(agentID)+".dna", "rb")
+            agentObj = pickle.load(tpfp)
+            tpfp.close()
+        else:
+            agentObj = sess.agentBasket[agentID]
+    '''        
     except Exception:
         print("Error couldnt retrieve error in getAgent()")
         agentObj = None
-    
+    '''
     return(agentObj)
 
 
-def generateAgentID(sessID):
+def generateAgentID(sess):
     """
     INPUT       : None
     OUTPUT      : Agent object
@@ -378,25 +405,32 @@ def generateAgentID(sessID):
     DESCRIPTION : Creates and stores the newly created AgentDna
                   object in that particular session's directory
     """
-    try:
-        if(not os.path.isdir(GA_UTIL_DIR+"/utilFiles")):
-            os.mkdir(GA_UTIL_DIR+"/utilFiles")
-        nextfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)+"/smtf/NEXTID.smtf", "r")
-        Idp = int(nextfp.readline())
-        nextfp.close()
-    except Exception:
-        Idp = 0
+    if(sess.mode == 'SAFE'):
+        try:
+            
+            if(not os.path.isdir(GA_UTIL_DIR+"/utilFiles")):
+                os.mkdir(GA_UTIL_DIR+"/utilFiles")
+            nextfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sess.sessID)+"/smtf/NEXTID.smtf", "r")
+            Idp = int(nextfp.readline())
+            nextfp.close()
+            
+        except Exception:
+            Idp = 0
 
-    try:
-        nextfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sessID)+"/smtf/NEXTID.smtf", "w")
-        nextfp.write(str(Idp+1))
-        nextfp.close()
-        return(Idp)
-
-    except Exception:
-        print("error in generate agent ID, couldnt write")
-        return(-1)
-
+        try:
+            nextfp = open(GA_UTIL_DIR+"/utilFiles/tmp"+str(sess.sessID)+"/smtf/NEXTID.smtf", "w")
+            nextfp.write(str(Idp+1))
+            nextfp.close()
+            return(Idp)
+        
+        except Exception:
+            print("error in generate agent ID, couldnt write")
+            return(-1)
+    else:
+        sess.nextAgentID += 1
+        return(sess.nextAgentID -1)
+    
+            
 
 def createAgent(sess):
     """
@@ -409,12 +443,12 @@ def createAgent(sess):
     
     agent = AgentClass.AgentDna(sess.numParam, sess.spread)
     agent.sessID = sess.sessID
-    agent.agentID = generateAgentID(sess.sessID)
+    agent.agentID = generateAgentID(sess)
 
-    storeAgent(agent)
+    storeAgent(sess, agent)
     return(agent)
 
-def deleteAgent(sessID, agentID):
+def deleteAgent(sess, agentID):
     """
     INPUT       : sessID, agentID
     OUTPUT      : None
@@ -423,25 +457,30 @@ def deleteAgent(sessID, agentID):
                   object in that particular session's directory
     """
     
-    lock(sessID)
-    try:
-        currGen = getCurrGen(sessID)
+    lock(sess)
+    if(1):
+        currGen = getCurrGen(sess)
         if(agentID in currGen):
-            toDelFile = GA_UTIL_DIR + ("/utilFiles/tmp" + str(sessID)
-                                   + "/dnaPool/dna" + str(agentID)
-                                   +".dna")
-            os.remove(toDelFile)
-            currGen.remove(agentID)
-            setCurrGen(sessID, currGen)
-            
+            if(sess.mode == 'SAFE'):
+                toDelFile = GA_UTIL_DIR + ("/utilFiles/tmp" + str(sess.sessID)
+                                           + "/dnaPool/dna" + str(agentID)
+                                           +".dna")
+                os.remove(toDelFile)
+                currGen.remove(agentID)
+                setCurrGen(sess, currGen)
+            else:
+                del sess.agentBasket[agentID]
+                currGen.remove(agentID)
+                setCurrGen(sess, currGen)
         else:
             print("Trying to delete an agent not present in the session")
-
+    '''
     except Exception:
         print("delete Agent encountered an error")
         return(0)
     
     finally:
-        unlock(sessID)
-        
+        unlock(sess.sessID)
+    '''
+    unlock(sess)
     return(1)
